Revert "poll sync on emitter initiation"
This reverts commit bf9cffd1d476a792ac9a4ca9e5244441724930e9.

+++ backend/src/main/java/com/sentinel/backend/repository/DataSyncRepository.java
@@ -20,7 +20,7 @@ public class DataSyncRepository {
         return jdbcTemplate.query(
                 DATA_SYNC_SQL,
                 (rs, rowNum) -> new DataSyncDTO(
-                        rs.getLong("signal_id"),
+                        rs.getInt("signal_id"),
                         rs.getString("question"),
                         (String[]) rs.getArray("options").getArray(),
                         rs.getString("status"),
